projects/catalyst/catalyst-agents/pyproject.toml
[project]
name = "catalyst-agents"
version = "0.0.1"
description = "Catalyst A2A agent runtime for OpenELIS Global"
authors = [
    {name = "OpenELIS Global Team", email = "team@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.115.2,<0.116.0",
    "uvicorn[standard]>=0.34.0,<0.35.0",
    "httpx>=0.28.1,<0.29.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "a2a-sdk[http-server]>=0.3.22,<0.4.0",
    "mcp>=1.1.2,<2.0.0",
    "google-genai>=0.2.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.23.8,<0.24.0",
    "honcho>=1.1.0,<2.0.0",
    "setuptools>=69.0.0,<79.0.0",
    "ruff>=0.8.0,<0.9.0",
    "mypy>=1.14.0,<2.0.0",
    "pre-commit>=4.0.0,<5.0.0",
    "types-requests>=2.32.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "TRY",    # tryceratops
    "RUF",    # ruff-specific rules
]
ignore = [
    "PLR0913",  # Too many arguments
    "TRY003",   # Avoid specifying long messages outside exception class
    "EM101",    # Exception must not use a string literal (style preference)
    "EM102",    # Exception must not use an f-string literal (style preference)
    "TID252",   # Prefer absolute imports (relative imports are fine in this structure)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "ARG001",  # Allow unused arguments in tests (fixtures)
    "ARG002",  # Allow unused method arguments in tests (protocol implementation)
    "PLR2004", # Allow magic values in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, tighten over time
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "a2a.*",
    "mcp.*",
    "google.genai.*",
]
ignore_missing_imports = true
